Update temperature_transfer.py
correct casting type

+++ temperature_transfer.py
@@ -1,3 +1,3 @@
-celsius_temp = int(input('請輸入攝氏溫度'))
+celsius_temp = float(input('請輸入攝氏溫度'))
 feihrenheit_temp = celsius_temp * 1.8 + 32
 print('轉換為華氏溫度是', feihrenheit_temp)
